Fixed bug causing multi_cell to reset position cursor to initial 'multi_cell' insert position, resulting in later cells & structures superimposing the multicell.
This code segment on line 1089 '0: lambda : self.set_xy(self.x + w, self.y),' is called when inserting a 'multi_cell' and is causing the problem with the cursor position when using a 'multi_cell', but does not seem to cause problems if commented out or removed.
This applies to the 'split_only' option, which also uses 'location_options' and seems to work fine and provide correct output whether line 1089 code is present or not.

The issue seems to be that the 'location' variable is not updated as 'multi_cell' runs with the intention to use it to reset the position cursor when 'split_only' is used.
However, it is being used even when 'split_only' is set to 'False', and resets the cursor position to the initial 'multi_cell' insert position.

In my limited testing, with Python 3.6.8, 3.7.7, and 3.8.3 the code change to line 1089 do not seem to produce negative effects to multi_cell or the return output when setting 'split_only' to 'True'.

Also, the 'split_only' option is very useful as you can use it to determine the line count/height of the multi_cell or divide out the cut lines for further use, but I cannot find any documentation or explanation regarding this option.

The 'multi_cell' section seems to need attention in general.  Another issue is that setting the fill flag to anything but '0' or 'False' ('Bool' or 'int') prevents the 'multi_cell' from being displayed.