drawing: remove global graphics state registry.
It was not my brightest idea to have this be global state, and it was
causing problems with writing the tests (result PDFs would be
different depending on what order the tests were run in, etc).
Honestly, having the registry be global didn't really make much sense
from a functionality perspective either, since it only really makes
sense to deduplicate the graphics state dictionaries across a single
document, not across all documents generated in a single python
invocation.

The scope of the graphics state dictionary registry is now constrained
to be per-FPDF instance, which should map much more reasonably to a
document-level scope, and importantly, also won't mess up the unit
tests.