drawing: fix some bugs in the debug render path.
At this point it almost seems like it may have been a good idea to
say "to hell with the performance hit" and unified the debug and
non-debug paths with a NullIO object that would just eat the
write calls and do nothing.

Though we'd still have to test both debug and non-debug behavior,
regardless, so it may not make the testing much more straightforward.