test.drawing: move test parameters to a separate file.
Since the parameters are just data, this is an experiment in separating
data from code to make each half more approachable.

Also quit using fixtures just for parameterizing tests in favor of
pytest.mark.parameterize. This allows the parameter arguments to be
passed in with individual names instead of needing to be destructured
from a single-name fixture.