New Line After Header *Before* Font Change
Currently, when exiting the h1/h2/etc tag the process was to first pop to the previous font and _then_ create a new line (`pdf.ln(self.h)`). The problem is that if the head font is large and the previous font was significantly smaller, then the new line is not large enough to escape past the bottom of the head font. I've simple reversed this order and tested it locally - seems to work perfectly now for various header and non-header font sizes.