svg: make path parsing follow the spec a bit more closely.
This prevents parsing of some erroneous paths, though some technically
incorrect paths can still be parsed okay. It does fix handling of some
previously missing features, such as the unary plus and chained
decimal separators acting to split numbers apart (e.g. "0.2.3" is
technically valid and should be parsed as "0.2 .3", though you'd have
to be some sort of criminal to actually use that in an SVG).