drawing: police GraphicsStyle attribute setting
This raises an AttributeError if a user tries to set a style attribute
that doesn't exist (for example, `style.dash_style` instead of
`style.stroke_dash_style`). I kept making typos during writing the
tests which would be silently swallowed and result in unexpected
output, which is exactly the behavior you don't want.