fpdf: use a pre-unscaled local context for inheriting document styles
To avoid double scaling of scale-sensitive inherited styles (right now
the only applicable ones being stroke width and dash pattern), we
create a local context and pre-unscale them. It would be nice not to
have to do this, but I can't think of a way around it that actually
works.

All of the drawing-based test PDFs will have to be regenerated.