drawing: decouple stroke_dash_pattern and stroke_dash_phase
Because these are coupled in the PDF serialization process, there's
some annoying jogging around that has to be done to make this work,
but it should hopefully make the API more intuitive to use. This also
makes some changes to the plumbing of the drawing.GraphicsStyle class
to make it more subclass friendly.