drawing: fix pylint nit
I honestly have no clue how pylint managed to propagate ellipsis to
this code, but miss that this if clause above specifically checks for
ellipsis and handles it separately. But also doesn't propagate it down
to the `render_list.append(intersection_rule.value)` call below.

Because this is a mystery, I hope this fix doesn't randomly fail on CI
again.