Allow add_link() to be called before add_page().
This patch sets the default page to 1 for add_link() if
no pages have yet been added to the PDF, otherwise the
default target is the current page.  Page 1 is mandatory,
so the target will definitely be valid even if it is not
later pointed elsewhere with set_link().

Since this is no longer an error, the test/test_links.py
test_inserting_link_with_no_page_number() has been changed
to ensure that it produces the same document as the later
test_later_call_to_set_link().