Make multipaged ToC work correctly without extra steps.
The general way to achieve this:
1. Create special text fragments so thay will be replaced with values when we know the values.
2. Update page numbers in links.

It's not necessary that every page has substitution fragments. A PDF file might have a lot of pages, so we want to avoid iterating through all of them. Thus it's better to map fragments to pages, not vice versa.

Page number in links are skewed after ToC moved pages. It should be updated as well.

The copy method of the Fragment class is an implementation of the Prototype pattern. It's useful here as we want to create new fragment based on an existing one.