Prevent merging fragments of different types.
An instance of `SubstitutionFragment` is also an instance of `Fragment`. With certain orderings, a substitution fragment could be extended with characters from another fragment. When the substitution fragment is later replaced with a value, all of those characters are lost.

Two consecutive substitution fragments will still be merged. Since there is only one substitution fragment, you'll see one value in the result file.